requirements.txt
boto3==1.38.9
fastapi==0.115.12
mysql-connector-python
python-dotenv==1.1.0
uvicorn[standard]
python-multipart